tests/user_service/unit_tests/requirements.txt
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
fastapi==0.104.1
httpx==0.25.2
pydantic[email]==2.5.0
passlib==1.7.4
python-jose==3.3.0
sqlalchemy==2.0.23
bcrypt==4.1.2
psycopg2-binary==2.9.9
aiokafka==0.10.0
email-validator==2.1.0
python-multipart==0.0.6